feat: Added address UI for both SS58 and H160 format
